requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
openai>=1.40.0
azure.identity
httpx>=0.24.0
python-dotenv>=1.0.0
